+++ react-playground/src/components/App/App.tsx
@@ -1,0 +1,12 @@
+import { Header } from "../Header";
+
+function App() {
+  return (
+    <div>
+      <Header />
+      Hello
+    </div>
+  );
+}
+
+export default App;
