--- react-playground/src/components/GeneApp.tsx
@@ -1,12 +1,0 @@
-import Header from "./Header";
-
-function GeneApp() {
-  return (
-    <div>
-      <Header />
-      Hello
-    </div>
-  );
-}
-
-export default GeneApp;
