+++ react-playground/src/components/Header/index.ts
@@ -1,0 +1,1 @@
+export { default as Header } from "./Header";
