+++ react-playground/src/index.tsx
@@ -1,12 +1,12 @@
 import React from "react";
 import ReactDOM from "react-dom";
 import "./index.css";
-import GeneApp from "./components/GeneApp";
+import { App } from "./components/App";
 import reportWebVitals from "./reportWebVitals";
 
 ReactDOM.render(
   <React.StrictMode>
-    <GeneApp />
+    <App />
   </React.StrictMode>,
   document.getElementById("root")
 );
