+++ react-playground/src/tests/App.test.tsx
@@ -1,7 +1,7 @@
 import { render, screen } from "@testing-library/react";
-import GeneApp from "../components/GeneApp";
+import { App } from "../components/App/";
 
 test("renders a message", () => {
-  render(<GeneApp />);
+  render(<App />);
   expect(screen.getByText("Hello")).toBeInTheDocument();
 });
