+++ react-playground/src/tests/Header.test.tsx
@@ -1,5 +1,5 @@
 import { fireEvent, render, screen } from "@testing-library/react";
-import Header from "../components/Header";
+import { Header } from "../components/Header";
 
 test("renders a message", () => {
   render(<Header />);
